scripts/requirements.txt
pillow==10.3.0
klayout==0.29.1
